requirements.txt
black>=22.6.0
coverage>=6.4.1
isort>=5.10.1
mypy>=0.961
pylint>=2.14.4
pytest>=7.1.2
pyaml==23.7.0
